refactor audio

+++ crates/common/src/geometry.rs
@@ -1,8 +1,8 @@
 /// Line between two points
-pub type Line = (mint::Point2<f32>, mint::Point2<f32>);
+pub type Line = (mint::Point2<f64>, mint::Point2<f64>);
 
 /// Rectangle from start of coordinates
 ///
-/// - start `mint::Point2<f32>`
-/// - size `mint::Vector2<f32>`
-pub type Rect = (mint::Point2<f32>, mint::Vector2<f32>);
+/// - start `mint::Point2<f64>`
+/// - size `mint::Vector2<f64>`
+pub type Rect = (mint::Point2<f64>, mint::Vector2<f64>);
