+++ crates/common/src/safe_area.rs
@@ -4,10 +4,10 @@ use serde::{Deserialize, Serialize};
 #[serde(rename_all = "camelCase")]
 /// Distances from screen edges to safe area
 pub struct SafeArea {
-    pub top: f32,
-    pub right: f32,
-    pub bottom: f32,
-    pub left: f32,
+    pub top: f64,
+    pub right: f64,
+    pub bottom: f64,
+    pub left: f64,
 }
 
 impl Default for SafeArea {
@@ -22,4 +22,4 @@ impl Default for SafeArea {
 }
 
 /// Default safe area insets
-pub const DEFAULT_SAFE_AREA: f32 = 18.0;
+pub const DEFAULT_SAFE_AREA: f64 = 18.0;
