+++ crates/common/src/test_util.rs
@@ -24,7 +24,7 @@ pub const TABLET_SCREEN_SIZES: [(u32, u32); 5] = [
 ];
 
 /// Safe area insets (pixels) for common mobile layouts (portrait).
-pub const MOBILE_SAFE_AREA_INSETS: [(f32, f32, f32, f32); 3] = [
+pub const MOBILE_SAFE_AREA_INSETS: [(f64, f64, f64, f64); 3] = [
     // (top, right, bottom, left)
     (44.0, 0.0, 34.0, 0.0), // iPhone X, 11, 12, 13, 14, 15 series w/ notch
     (24.0, 0.0, 16.0, 0.0), // Typical Android (navigation bar, status bar)
@@ -32,12 +32,12 @@ pub const MOBILE_SAFE_AREA_INSETS: [(f32, f32, f32, f32); 3] = [
 ];
 
 /// Safe area insets (pixels) for common tablet layouts (portrait).
-pub const TABLET_SAFE_AREA_INSETS: [(f32, f32, f32, f32); 2] = [
+pub const TABLET_SAFE_AREA_INSETS: [(f64, f64, f64, f64); 2] = [
     (24.0, 0.0, 20.0, 0.0), // Typical iPad Pro with home indicator area
     (0.0, 0.0, 0.0, 0.0),   // Older iPads, Android tablets, no system gesture zone
 ];
 
-pub fn test_cases() -> impl Iterator<Item = (&'static (u32, u32), &'static (f32, f32, f32, f32))> {
+pub fn test_cases() -> impl Iterator<Item = (&'static (u32, u32), &'static (f64, f64, f64, f64))> {
     DESKTOP_SCREEN_SIZES
         .iter()
         .zip(DESKTOP_SCREEN_SIZES.iter().map(|_| {
