+++ crates/frontend/src/components/intro/consts.rs
@@ -2,20 +2,20 @@
 // Intro artwork reference (moved from layout; kept front & center here)
 // -----------------------------------------------------------------------------
 
-pub const INTRO_COMP_VIEWBOX_WIDTH: f32 = 430.0;
-pub const INTRO_COMP_VIEWBOX_HEIGHT: f32 = 932.0;
+pub const INTRO_COMP_VIEWBOX_WIDTH: f64 = 430.0;
+pub const INTRO_COMP_VIEWBOX_HEIGHT: f64 = 932.0;
 
 /// Sun radius within artwork space
-pub const INTRO_SUN_RADIUS: f32 = 39.0;
+pub const INTRO_SUN_RADIUS: f64 = 39.0;
 /// Sun center X within artwork space
-pub const INTRO_SUN_POS_X: f32 = 107.0;
+pub const INTRO_SUN_POS_X: f64 = 107.0;
 /// Sun center Y within artwork space
-pub const INTRO_SUN_POS_Y: f32 = 164.0;
+pub const INTRO_SUN_POS_Y: f64 = 164.0;
 /// Flute center X within artwork space
-pub const INTRO_FLUTE_POS_X: f32 = 48.3365;
+pub const INTRO_FLUTE_POS_X: f64 = 48.3365;
 /// Flute center Y within artwork space
-pub const INTRO_FLUTE_POS_Y: f32 = 585.964;
+pub const INTRO_FLUTE_POS_Y: f64 = 585.964;
 /// Flute rotation (degrees)
-pub const INTRO_FLUTE_ROT_DEG: f32 = -17.1246;
+pub const INTRO_FLUTE_ROT_DEG: f64 = -17.1246;
 
 pub const INTRO_FADE_DURATION_MS: f64 = 2000.0;
